refactor: migrate web app to standard Next.js src/ layout
Move app/ and srcs/ into src/ following the convention established in
vercel/next.js#8451 and now default in create-next-app. Merges
srcs/pages/ into src/components/ since those files are shared
components, not routes. Updates all @/srcs/* import paths accordingly.
