+++ apps/web/src/app/admin/login/page.tsx
@@ -1,4 +1,4 @@
-import AdminLoginPage from '@/srcs/pages/admin_login_page';
+import AdminLoginPage from '@/src/components/admin_login_page';
 
 export default function AdminLoginRoute() {
   return <AdminLoginPage />;
