+++ apps/web/src/app/favicon.ico
