+++ apps/web/src/components/decorative_birds.tsx
