+++ apps/web/src/components/photo_gallery.tsx
